fix: move docker base dir creation to fact_base
* because it is also used by the PDF report container in the frontend
